{{ cookiecutter.project_slug }}/requirements_Dev.txt
amqp==5.0.6
asgiref==3.4.1
autopep8==1.5.7
backports.entry-points-selectable==1.1.0
billiard==3.6.4.0
black==21.10b0
celery==5.1.2
certifi==2021.10.8
cffi==1.15.0
cfgv==3.3.1
charset-normalizer==2.0.7
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
defusedxml==0.7.1
distlib==0.3.3
dj-rest-auth==2.1.11
Django==3.2.8
django-allauth==0.44.0
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-configurations==2.2
django-cors-headers==3.10.0
django-debug-toolbar==3.2.2
django-environ==0.8.1
django-extensions==3.1.3
django-filter==21.1
django-timezone-field==4.2.1
djangorestframework==3.12.4
dnspython==2.1.0
drf-psq==1.1.0
drf-yasg==1.20.0
filelock==3.3.1
identify==2.3.0
idna==3.3
inflection==0.5.1
isort==5.9.3
itypes==1.2.0
Jinja2==3.0.2
kombu==5.1.0
MarkupSafe==2.0.1
mypy-extensions==0.4.3
nodeenv==1.6.0
oauthlib==3.1.1
packaging==21.0
pathspec==0.9.0
Pillow==8.4.0
platformdirs==2.4.0
pre-commit==2.15.0
prompt-toolkit==3.0.21
psycopg2==2.9.2
pycodestyle==2.8.0
pycparser==2.20
Pygments==2.10.0
PyJWT==2.3.0
pyparsing==2.4.7
python-crontab==2.6.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
PyYAML==6.0
regex==2021.11.2
requests==2.26.0
requests-oauthlib==1.3.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==1.2.2
typing-extensions==3.10.0.2
uritemplate==4.1.1
urllib3==1.26.7
uwsgi==2.0.20
vine==5.0.0
virtualenv==20.8.1
wcwidth==0.2.5
